+++ persistent/Database/Persist/EntityDef.hs
@@ -19,6 +19,7 @@ module Database.Persist.EntityDef
     , getEntityKeyFields
     , getEntityComments
     , getEntityExtra
+    , getEntitySchema
     , isEntitySum
     , entityPrimary
     , entitiesPrimary
@@ -27,6 +28,7 @@ module Database.Persist.EntityDef
     , setEntityId
     , setEntityIdDef
     , setEntityDBName
+    , setEntitySchema
     , overEntityFields
       -- * Related Types
     , EntityIdDef(..)
@@ -88,6 +90,13 @@ getEntityDBName = entityDB
 
 getEntityExtra :: EntityDef -> Map Text [[Text]]
 getEntityExtra = entityExtra
+
+-- | Get the (optional) schema for an entity. Specified with the syntax
+-- `schema=foo` on the Persistent model.
+--
+-- @since 2.15.0
+getEntitySchema :: EntityDef -> Maybe SchemaNameDB
+getEntitySchema = entitySchema
 
 -- |
 --
@@ -194,6 +203,12 @@ getEntityKeyFields = entityKeyFields
 -- @since 2.13.0.0
 setEntityFields :: [FieldDef] -> EntityDef -> EntityDef
 setEntityFields fd ed = ed { entityFields = fd }
+
+-- | Update the entity schema.
+--
+-- @since 2.15.0
+setEntitySchema :: Maybe SchemaNameDB -> EntityDef -> EntityDef
+setEntitySchema sn ed = ed { entitySchema = sn }
 
 -- | Perform a mapping function over all of the entity fields, as determined by
 -- 'getEntityFieldsDatabase'.