+++ persistent/test/Database/Persist/THSpec.hs
@@ -78,7 +78,7 @@ type TextId = Text
 
 share [mkPersistWith  sqlSettings { mpsGeneric = False, mpsDeriveInstances = [''Generic] } [entityDef @JsonEncodingSpec.JsonEncoding Proxy]] [persistUpperCase|
 
-Person json
+Person json schema=some_schema
     name Text
     age Int Maybe
     foo Foo
@@ -371,6 +371,7 @@ spec = describe "THSpec" $ do
                             , entityExtra = mempty
                             , entitySum = False
                             , entityComments = Nothing
+                            , entitySchema = Nothing
                             }
         it "has the cascade on the field def" $ do
             fieldCascade subject `shouldBe` expected
@@ -506,6 +507,13 @@ spec = describe "THSpec" $ do
         it "has a good safe to insert class instance" $ do
             let proxy = Proxy :: SafeToInsert CustomIdName => Proxy CustomIdName
             proxy `shouldBe` Proxy
+    describe "Entity Schema" $ do
+        let personDef =
+                entityDef (Proxy :: Proxy Person)
+        it "reads the entity schema" $ do
+            (entitySchema personDef)
+                `shouldBe`
+                    (Just $ SchemaNameDB "some_schema")
 
 (&) :: a -> (a -> b) -> b
 x & f = f x