+++ persistent-qq/ChangeLog.md
@@ -1,5 +1,9 @@
 # Changelog for persistent-qq
 
+## 2.12.0.7
+
+* Add tests for the new `schema=` annotation. [#1561](https://github.com/yesodweb/persistent/pull/1561)
+
 ## 2.12.0.6
 
 * Fix test compilation by importing `Control.Monad` explicitly [#1487](https://github.com/yesodweb/persistent/pull/1487)