+++ persistent-sqlite/ChangeLog.md
@@ -1,5 +1,11 @@
 # Changelog for persistent-sqlite
 
+## 2.13.4
+
+* [#1561](https://github.com/yesodweb/persistent/pull/1561)
+    * Add schema support to Persistent.
+    * Add 'attachDatabaseFile' for attaching database files as schemas.
+
 ## 2.13.3.0
 
 * [#1524](https://github.com/yesodweb/persistent/pull/1524)