+++ persistent-test/ChangeLog.md
@@ -1,4 +1,7 @@
-## Unreleased changes
+## 2.13.2
+
+* [#1561](https://github.com/yesodweb/persistent/pull/1561)
+    * Add schema support to Persistent.
 
 ## 2.13.1.3